changes ansible style to playbook
This commit changes the complete structure according to
http://docs.ansible.com/playbooks_best_practices.html#directory-layout

The mailserver to build is now in the group "mailservers" and belongs to
the roles "common" and "mailserver". Tasks, variables, handlers, files
and templates for the roles are now defined in "roles".

Legacy style ${var} have largely been replaced by {{ var }}.

There is a single configuration file for private data in
ansible/host_vars which has to be filled with mysql connection
parameters. The file "mail.localhost" would build a mailerver with that
hostname, exclusively - so this filename needs to be changed.
Accordingly, the new hostname has to be given in ansible/hosts.

Dockerfiles was modified to call site.yml