Further fixes.
This commit changes the reject lists to use spamhaus' zen list. Also,
further configuration changes to postfix are made.

The role "common" is removed, we are in a single-instance setup here. No
further roles necessary.

Since the mail direcotry is an exposed volume, docker mounts it's
contents as root. Therefore, the startup-script will chown all
directorys under $VMAIL_DIR with 770 and all files with 660.